requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
supabase
gunicorn
python-jose[cryptography] 